#63. Fix minor compilation errors.

+++ Executable/Platformer2dDemo/Logic/Platformer/Spawn/Spawn.cpp
@@ -160,7 +160,7 @@ void SpawnManager::ProcessScheduledSpawns (const Emergence::Celerity::TimeSingle
             layerSetup->layer = layer.value ();
         }
 
-        Emergence::Container::Optional<uint16_t> team;
+        Emergence::Container::Optional<uint8_t> team;
         if (auto teamCursor = fetchTeamByObjectId.Execute (&spawn->objectId);
             const auto *spawnTeam = static_cast<const TeamComponent *> (*teamCursor))
         {
